refactor(analyse): move measurement functions
Moved measurement functions from commands/analyse/analyse.py to
commands/analyse/utils/measurement.py.
- Updated analyse import statements.
