+++ src/evaluator/commands/analyse/analyse.py
@@ -150,20 +150,20 @@ def processComponent(component_label, labelled_volumes, component_properties, vo
     scale = voxel_size_nm if voxel_size_nm is not None else 1.0
     scale_label = "nm" if voxel_size_nm is not None else "vox"
     lg.debug(f"analyse | {filename} | Component {component_label} | Creating component mask...")
-    component_mask = createComponentMask(component=component_properties, labelled_vol=labelled_volumes, label_val=component_label)
+    component_mask = measurement.createComponentMask(component=component_properties, labelled_vol=labelled_volumes, label_val=component_label)
     lg.debug(f"analyse | {filename} | Component {component_label} | Measuring membrane volume and equivalent diameter...")
-    membrane_vol_nm3, equiv_diameter_nm = measureMembraneVolumeDiameter(component=component_properties, scale=scale)
+    membrane_vol_nm3, equiv_diameter_nm = measurement.measureMembraneVolumeDiameter(component=component_properties, scale=scale)
     component_mask_dilated = filtering.morphologicalDilation(component_mask)
     lg.debug(f"analyse | {filename} | Component {component_label} | Checking if component is enclosed...")
     enclosed, fill_ratio = filtering.checkEnclosed(component_mask=component_mask_dilated, threshold=fill_threshold)
     lg.debug(f"analyse | {filename} | Component {component_label} | Measuring lumen volume...")
-    lumen_vol_nm3 = measureLumenVolume(component_mask=component_mask, scale=scale)
+    lumen_vol_nm3 = measurement.measureLumenVolume(component_mask=component_mask, scale=scale)
     lg.debug(f"analyse | {filename} | Component {component_label} | Measuring surface area...")
-    surface_area = computeSurfaceArea(component_mask, voxel_size_nm)
+    surface_area = measurement.computeSurfaceArea(component_mask, voxel_size_nm)
     lg.debug(f"analyse | {filename} | Component {component_label} | Measuring major/minor axes diameters...")
     major_axis_diameter, minor_axis_diameter = geometry.measureAxes(component=component_properties, equiv_diameter_nm=equiv_diameter_nm)
     lg.debug(f"analyse | {filename} | Component {component_label} | Measuring eccentricity and aspect ratio...")
-    eccentricity, aspect_ratio = measureEccentricityAspectRatio(major_axis_diameter=major_axis_diameter, minor_axis_diameter=minor_axis_diameter)
+    eccentricity, aspect_ratio = measurement.measureEccentricityAspectRatio(major_axis_diameter=major_axis_diameter, minor_axis_diameter=minor_axis_diameter)
     return {
         "tomogram": filename,
         "label": component_label,
@@ -181,8 +181,6 @@ def processComponent(component_label, labelled_volumes, component_properties, vo
         "measurement_units": scale_label,
     }
 
-
-
 # =========================
 # DEFINE FUNCTION: morphologicalClosure
 # =========================
@@ -193,82 +191,6 @@ def morphologicalClosure(binary_vol: numpy.ndarray):
     '''
     struc = ndimage.generate_binary_structure(3, 1)
     return ndimage.binary_closing(binary_vol, structure=struc, border_value=False)
-
-
-# =========================
-# DEFINE FUNCTION: computeSurfaceArea
-# =========================
-def computeSurfaceArea(component_mask: numpy.ndarray, voxel_size_nm: float):
-    '''
-    Estimates surface area using the marching cubes algorithm.
-    Returns nm^2 if voxel size is known, otherwise vox^2.
-    '''
-    try:
-        verts, faces, _, _ = measure.marching_cubes(
-            component_mask.astype(numpy.uint8),
-            level=0.5,
-            spacing=(1.0, 1.0, 1.0),
-        )
-        sa_vox = measure.mesh_surface_area(verts, faces)
-    except (ValueError, RuntimeError):
-        return numpy.nan
-    if voxel_size_nm is not None:
-        return sa_vox * (voxel_size_nm ** 2)
-    return sa_vox
-
-# =========================
-# DEFINE FUNCTION: measureMembraneVolumeDiameter
-# =========================
-def measureMembraneVolumeDiameter(component, scale):
-    '''
-    Measures the volume of membrane components in voxels and converts to nm^3.
-    Calculates the equivalent spherical diameter in nm.
-    '''
-    vol_vox = component.area
-    vol_nm3 = vol_vox * (scale ** 3)
-    equiv_diameter_nm = (6 * vol_nm3 / numpy.pi) ** (1 / 3)
-    return vol_nm3, equiv_diameter_nm
-
-# =========================
-# DEFINE FUNCTION: createComponentMask
-# =========================
-def createComponentMask(component, labelled_vol, label_val):
-    '''
-    Extract the bounding-box sub-volume for a given component label and return
-    a boolean mask.
-    '''
-    bbox = component.bbox
-    slices = (
-        slice(bbox[0], bbox[3]),
-        slice(bbox[1], bbox[4]),
-        slice(bbox[2], bbox[5]),
-    )
-    return labelled_vol[slices] == label_val
-
-# =========================
-# DEFINE FUNCTION: measureLumenVolume
-# =========================
-def measureLumenVolume(component_mask, scale):
-    '''
-    Calculate the lumen volume by filling holes and subtracting the membrane shell.
-    '''
-    filled_mask = ndimage.binary_fill_holes(component_mask)
-    lumen_vol_vox = numpy.sum(filled_mask) - numpy.sum(component_mask)
-    return lumen_vol_vox * (scale ** 3)
-
-# =========================
-# DEFINE FUNCTION: measureEccentricityAspectRatio
-# =========================
-def measureEccentricityAspectRatio(major_axis_diameter, minor_axis_diameter):
-    '''
-    Given major and minor diameters, calculate eccentricity and aspect ratio.
-    Eccentricity: 0 (spherical) → 1 (tubular).
-    '''
-    semimajor = major_axis_diameter / 2
-    semiminor = minor_axis_diameter / 2
-    eccentricity = numpy.sqrt(1 - (semiminor / semimajor) ** 2) if semimajor > 0 else numpy.nan
-    aspect_ratio = major_axis_diameter / minor_axis_diameter if minor_axis_diameter > 0 else numpy.nan
-    return eccentricity, aspect_ratio
 
 # =========================
 # DEFINE FUNCTION: saveResultsCSV
